requirements.txt
aiogram==3.24.0
asyncpg==0.31.0
redis==7.1.0
yt-dlp==2025.12.8
python-dotenv==1.2.1
aiofiles==25.1.0